refactor(sputnik): move wasi polyfill lifecycle to module
